feat(users): add delete user route
