+++ src/models/userModel.js
@@ -44,6 +44,11 @@ const userSchema = new mongoose.Schema({
   passwordChangedAt: Date,
   passwordResetToken: String,
   passwordResetExpires: Date,
+  active: {
+    type: Boolean,
+    default: true,
+    select: false,
+  },
 });
 
 userSchema.pre('save', async function (next) {
@@ -93,6 +98,12 @@ userSchema.pre('save', function (next) {
   this.passwordChangedAt = Date.now() - 1000;
 
   next();
+});
+
+userSchema.pre(/^find/, function (next) {
+  this.find({ active: { $ne: false } });
+
+  return next();
 });
 
 const User = mongoose.model('User', userSchema);
